refactor: extract link component
