--- src/util/links.ts
@@ -1,8 +1,0 @@
-export const LINKS = [
-    { name: 'Email', link: 'mailto:contact@csclub.org.au' },
-    { name: 'GitHub', link: 'https://github.com/compsci-adl' },
-    { name: 'Instagram', link: 'https://www.instagram.com/compsci.adl/' },
-    { name: 'Facebook', link: 'https://www.facebook.com/compsci.adl/' },
-    { name: 'Discord', link: 'https://discord.gg/UjvVxHA' },
-    { name: 'LinkedIn', link: 'https://www.linkedin.com/company/compsci-adl/' },
-] as const;
